[refactor] #54 dto 수정, if문 수정
