+++ src/main/java/com/example/eatmate/app/domain/chatRoom/dto/response/ChatRoomResponseDto.java
@@ -5,6 +5,8 @@ import java.util.List;
 import org.springframework.data.domain.Page;
 
 import com.example.eatmate.app.domain.chat.dto.response.ChatMessageResponseDto;
+import com.example.eatmate.app.domain.member.domain.Mbti;
+import com.example.eatmate.app.domain.member.domain.Member;
 
 import lombok.Builder;
 import lombok.Getter;
@@ -13,7 +15,7 @@ import lombok.NoArgsConstructor;
 @Getter
 @NoArgsConstructor
 public class ChatRoomResponseDto {
-	private Page<ChatMessageResponseDto> chats;
+	private List<ChatMessageResponseDto> chats;
 	private List<ChatMemberResponseDto> participants;
 	private ChatRoomDeliveryNoticeDto deliveryNotice;
 	private ChatRoomOfflineNoticeDto offlineNotice;
@@ -23,7 +25,7 @@ public class ChatRoomResponseDto {
 	@Builder
 	private ChatRoomResponseDto(Page<ChatMessageResponseDto> chats, List<ChatMemberResponseDto> participants,
 		ChatRoomDeliveryNoticeDto deliveryNotice, ChatRoomOfflineNoticeDto offlineNotice) {
-		this.chats = chats;
+		this.chats = chats.getContent();
 		this.participants = participants;
 		this.deliveryNotice = deliveryNotice;
 		this.offlineNotice = offlineNotice;
@@ -45,5 +47,24 @@ public class ChatRoomResponseDto {
 			.participants(participants)
 			.offlineNotice(offlineNotice)
 			.build();
+	}
+
+	@Getter
+	public static class ChatMemberResponseDto {
+		private String memberName;
+		private Mbti mbti;
+
+		@Builder
+		private ChatMemberResponseDto(String memberName, Mbti mbti) {
+			this.memberName = memberName;
+			this.mbti = mbti;
+		}
+
+		public static ChatMemberResponseDto from(Member member) {
+			return ChatMemberResponseDto.builder()
+				.memberName(member.getName())
+				.mbti(member.getMbti())
+				.build();
+		}
 	}
 }
